+++ RestMan.Context/Models/Order.cs
@@ -9,14 +9,14 @@ namespace RestMan.Context.Models
 {
     public class Order
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         [Required]
         public DateTime CreatedAt { get; set; } = DateTime.Now;
-        public int WaiterId { get; set; }
+        public long WaiterId { get; set; }
         public User Waiter { get; set; }
-        public int TableId { get; set; }
+        public long TableId { get; set; }
         public Table Table { get; set; }
-        public int ShiftId { get; set; }
+        public long ShiftId { get; set; }
         public Shift Shift { get; set; }
         public DateTime? DeletedAt { get; set; }
         public int? PaidByGiftCard { get; set; } = 0;
